+++ comments.py
@@ -2,21 +2,31 @@ from fastapi import APIRouter
 from fastapi.encoders import jsonable_encoder
 from schemas import AddressModel, PlacesModel, TravelCategoryModel, CommentsModel, TravelsModel
 from db.database import Session, ENGINE
-from fastapi import HTTPException, status
+from fastapi import HTTPException, status, Depends
 from db.models import Address, TravelCategory, Travels, Comments, Users, Places
+from fastapi.encoders import jsonable_encoder
+from fastapi_jwt_auth import AuthJWT
 
 comments_router = APIRouter(prefix="/comments")
 session = Session(bind=ENGINE)
 
 
 @comments_router.get("/")
-async def comments_get():
+async def comments_get(Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
     comments = session.query(Comments).all()
     return comments
 
 
 @comments_router.post("/")
-async def create_comment(comments: CommentsModel):
+async def create_comment(comments: CommentsModel, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
     exist_user = session.query(Users).filter(Users.username == comments.user).first()
     if exist_user:
         new_comment = Comments(
@@ -32,7 +42,11 @@ async def create_comment(comments: CommentsModel):
 
 
 @comments_router.get("/{id}")
-async def read_comment(id: int):
+async def read_comment(id: int, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
     comment = session.query(Comments).filter(Comments.id == id).first()
     if comment:
         return HTTPException(status_code=status.HTTP_200_OK, detail=comment)
@@ -41,7 +55,11 @@ async def read_comment(id: int):
 
 
 @comments_router.put("/{id}")
-async def update_comment(id: int, comment: CommentsModel):
+async def update_comment(id: int, comment: CommentsModel, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
     commentw = session.query(Comments).filter(Comments.id == id).first()
     if not commentw:
         return HTTPException(status_code=status.HTTP_404_NOT_FOUND)
@@ -56,17 +74,31 @@ async def update_comment(id: int, comment: CommentsModel):
 
 
 @comments_router.delete("/{id}")
-async def delete_comment(id: int):
-    comment = session.query(Comments).filter(Comments.id == id).first()
-    if not comment:
-        return HTTPException(status_code=status.HTTP_404_NOT_FOUND)
-    session.delete(comment)
-    session.commit()
-    return HTTPException(status_code=status.HTTP_204_NO_CONTENT)
+async def delete_comment(id: int, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
+    exist_user = session.query(Users).filter(Users.username == Authentization.get_jwt_subject()).first()
+    if not exist_user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="The user does not exist")
+    if exist_user.is_staff:
+        comment = session.query(Comments).filter(Comments.id == id).first()
+        if not comment:
+            return HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+        session.delete(comment)
+        session.commit()
+        return HTTPException(status_code=status.HTTP_204_NO_CONTENT)
+    else:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="It is not possible to view user data for you")
 
 
 @comments_router.get("/{id}/about_comment")
-async def active_users(id: int):
+async def active_users(id: int, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
     comments = session.query(Comments).filter(Comments.id == id).first()
     if not comments:
         return HTTPException(status_code=status.HTTP_404_NOT_FOUND)
