+++ travel_category.py
@@ -1,43 +1,63 @@
 from fastapi import APIRouter
 from schemas import AddressModel, PlacesModel, TravelCategoryModel, CommentsModel, TravelsModel, CheckUserModel
 from db.database import Session, ENGINE
-from fastapi import HTTPException, status
+from fastapi import HTTPException, status, Depends
 from db.models import Address, TravelCategory, Travels, Comments, Users, Places
 from fastapi.encoders import jsonable_encoder
+from fastapi_jwt_auth import AuthJWT
 
 session = Session(bind=ENGINE)
 travel_category_router = APIRouter(prefix="/travelcategory")
 
 
 @travel_category_router.get("/")
-async def travel_categories_get():
+async def travel_categories_get(Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
     travel_categories = session.query(TravelCategory).all()
     return travel_categories
 
 
 @travel_category_router.post("/")
-async def create_travel_category(travel_category: TravelCategoryModel):
-    exist_author = session.query(Users).filter(Users.username == travel_category.author).first()
-    if exist_author is None:
-        return HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User not found")
+async def create_travel_category(travel_category: TravelCategoryModel, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
+    exist_user = session.query(Users).filter(Users.username == Authentization.get_jwt_subject()).first()
+    if not exist_user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="The user does not exist")
+    if exist_user.is_staff:
+        exist_author = session.query(Users).filter(Users.username == travel_category.author).first()
+        if exist_author is None:
+            return HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User not found")
+        else:
+            author_id = exist_author.id
+        exist_travel_category = session.query(TravelCategory).filter(
+            TravelCategory.name == travel_category.name).first()
+        if exist_travel_category:
+            return HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Travelcategory already exists")
+        else:
+            new_category = TravelCategory(
+                name=travel_category.name,
+                author_id=author_id
+            )
+            session.add(new_category)
+            session.commit()
+            session.refresh(new_category)
+            return {"message": "TravelCategory successful created", "travel_category": new_category}
     else:
-        author_id = exist_author.id
-    exist_travel_category = session.query(TravelCategory).filter(TravelCategory.name == travel_category.name).first()
-    if exist_travel_category:
-        return HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Travelcategory already exists")
-    else:
-        new_category = TravelCategory(
-            name=travel_category.name,
-            author_id=author_id
-        )
-        session.add(new_category)
-        session.commit()
-        session.refresh(new_category)
-        return {"message": "TravelCategory successful created", "travel_category": new_category}
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="It is not possible to view user data for you")
 
 
 @travel_category_router.get("/{id}")
-def get_travel_category(id: int):
+def get_travel_category(id: int, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
     exist_travel_category = session.query(TravelCategory).filter(TravelCategory.id == id).first()
     if not exist_travel_category:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
@@ -45,36 +65,57 @@ def get_travel_category(id: int):
 
 
 @travel_category_router.put("/{id}")
-def update_travel_category(id: int, travel_category: TravelCategoryModel, user: CheckUserModel):
-    user = session.query(Users).filter(Users.id == id).first()
-    if not user:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User not found")
-    if user.is_staff is False:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="This user is not staff")
-    exist_travel_category = session.query(TravelCategory).filter(TravelCategory.id == id).first()
-    if not exist_travel_category:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
-    author = session.query(Users).filter(Users.username == travel_category.author).first()
-    if author is None:
-        return HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User not found")
-    exist_travel_category.author_id = author.id
-    exist_travel_category.name = travel_category.name
-    session.commit()
-    return HTTPException(status_code=status.HTTP_200_OK, detail="successfully updated")
+def update_travel_category(id: int, travel_category: TravelCategoryModel, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
+    exist_user = session.query(Users).filter(Users.username == Authentization.get_jwt_subject()).first()
+    if not exist_user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="The user does not exist")
+    if exist_user.is_staff:
+        exist_travel_category = session.query(TravelCategory).filter(TravelCategory.id == id).first()
+        if not exist_travel_category:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+        author = session.query(Users).filter(Users.username == travel_category.author).first()
+        if author is None:
+            return HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User not found")
+        exist_travel_category.author_id = author.id
+        exist_travel_category.name = travel_category.name
+        session.commit()
+        return HTTPException(status_code=status.HTTP_200_OK, detail="successfully updated")
+    else:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="It is not possible to view user data for you")
 
 
 @travel_category_router.delete("/{id}")
-def delete_travel_category(id: int):
-    exist_travel_category = session.query(TravelCategory).filter(TravelCategory.id == id).first()
-    if not exist_travel_category:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
-    session.delete(exist_travel_category)
-    session.commit()
-    return HTTPException(status_code=status.HTTP_204_NO_CONTENT)
+def delete_travel_category(id: int, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
+    exist_user = session.query(Users).filter(Users.username == Authentization.get_jwt_subject()).first()
+    if not exist_user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="The user does not exist")
+    if exist_user.is_staff:
+        exist_travel_category = session.query(TravelCategory).filter(TravelCategory.id == id).first()
+        if not exist_travel_category:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+        session.delete(exist_travel_category)
+        session.commit()
+        return HTTPException(status_code=status.HTTP_204_NO_CONTENT)
+    else:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="It is not possible to view user data for you")
+
+
 
 
 @travel_category_router.get("/{id}/author")
-async def get_travel_category_author(id: int):
+async def get_travel_category_author(id: int, Authentization: AuthJWT = Depends()):
+    try:
+        Authentization.jwt_required()
+    except:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token is invalid")
     exist_travel_category = session.query(TravelCategory).filter(TravelCategory.id == id).first()
     if not exist_travel_category:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
